Extract repeated <ErrorBanner> parts of staging and pathogens into component [#1133]
